Vue Router and initial Run Start event (#3)
* Implement vue-router

* properly test that component loads on route

* Add route for run view

* Initial load of RunViewer

* Fix issues from localStorage not being tested

* Test localStorageDriver.create

* One last test on create

* Test localStorageDriver.retrieve

* Remove todo

* Implement initial run event and navigate to new run

* Update todo

* Display run start event with some styling

* Little more style, add select starter button

* Update readme

+++ src/services/api/localStorageDriver.js
@@ -1,17 +1,46 @@
-export function create (keys, data){
-  const path = buildPathPartsFromKeys(keys).join('/')
-  //get the current list data
-  const currentList = getList(path)
-  //create the individual item
-  localStorage.setItem(
-    `${path}/${currentList.nextId}`, 
-    JSON.stringify(Object.assign(data, {id: currentList.nextId}))
-  )
-  //update the list
-  currentList.childKeys.push(currentList.nextId)
-  currentList.nextId++
-  delete currentList.children
-  localStorage.setItem(path, JSON.stringify(currentList))
+const isDefined = (val) => {
+  return val !== undefined && val !== null
+}
+
+/**
+ * 
+ * @param {Array<Object>} keys - parent keys of the item that should be inserted, signature [{ key: String, [value]: String }]
+ * @param {Object} data - Data to insert at the location indicated by keys
+ */
+export function create (keys, data, options={}){
+  try{
+    const lastKeyValue = keys.slice(-1)[0].value
+    const pathParts = buildPathPartsFromKeys(keys)
+    let listPath
+    if( isDefined(lastKeyValue) ){
+      //get the current list data
+      listPath = pathParts.slice(0,-1).join('/')
+    }
+    else {
+      listPath = pathParts.join('/')
+    }
+    const currentList = getShallowList(listPath)
+    const nextId = isDefined(lastKeyValue) ? lastKeyValue : currentList.nextId
+    if( currentList.childKeys.find(key => key === nextId) ) throw new Error('Item already exists')
+    localStorage.setItem(
+      `${listPath}/${nextId}`, 
+      JSON.stringify(Object.assign(data, {id: nextId}))
+    )
+    //update the list
+    currentList.childKeys.push(nextId)
+    currentList.childKeys.sort()
+    currentList.nextId = nextId >= currentList.nextId ? nextId+1 : currentList.nextId
+    localStorage.setItem(listPath, JSON.stringify(currentList))
+    if( options.returnCreated ){
+      return {
+        path: `/${listPath}/${nextId}`,
+        item: data
+      }
+    }
+  } catch (e) {
+    e.code = 500
+    throw e
+  }
 }
 
 export function update (keys, data){}
@@ -21,20 +50,46 @@ export function replace (keys, data){}
 export function remove (keys, data){}
 
 export function retrieve (keys){
-  const pathParts = buildPathPartsFromKeys(keys)
   let returnValue
-  if( pathParts.length !== keys.length*2 ){
-    //we're getting a list of items, and we assume the last key had a value of null
-    const path = pathParts.join('/')
-    returnValue = getList(path).children
+  try{
+    const pathParts = buildPathPartsFromKeys(keys)
+    let path
+    if( pathParts.length !== keys.length*2 ){
+      //we're getting a list of items, and we assume the last key had a value of null
+      path = pathParts.join('/')
+      returnValue = getList(path).children
+    }
+    else{
+      //we're getting a specific item, just retrieve it
+      path = pathParts.join('/')
+      //this will throw if access is denied
+      returnValue = JSON.parse(localStorage.getItem(path))
+    }
+  } catch (e){
+    e.code = 500
+    throw e
   }
-  else{
-    //we're getting a specific item, just retrieve it
-    path = pathParts.join('/')
-    //this will throw if access is denied
-    returnValue = JSON.parse(localStorage.getItem(path))
+
+  if( !returnValue ) {
+    const notFound = new Error('Item not found')
+    notFound.code = 404
+    throw notFound
   }
   return returnValue
+}
+
+function getShallowList(path){
+  const listData = JSON.parse(localStorage.getItem(path))
+  if( !listData ) {
+    return {
+      nextId: 0,
+      childKeys: []
+    }
+  }
+  return {
+    nextId: listData.nextId,
+    childKeys: listData.childKeys
+  }
 }
 
 function getList(path){
@@ -62,7 +117,9 @@ function getList(path){
 }
 
 function buildPathPartsFromKeys(keys){
-  return keys.map((key) => {
-    return `${encodeURIComponent(key.key)}${key.value ? `/${key.value}` : ''}`
-  })
+  return keys.reduce((parts, key) => {
+    parts.push(encodeURIComponent(key.key))
+    if( key.value !== undefined && key.value !== null ) parts.push(`${key.value}`)
+    return parts
+  }, [])
 }
